Components
Integrated the two systems. A component based damage system is now tested and fully functional.  TODO is design a text format for making components outside the editor for brevity sake. Testing systems are currently in place for creating default templates and automatically implementing them, but these need to be removed and replaced with more permanent systems.